Refactor post.php template and add conditional rendering for admin actions

+++ templates/post.php
@@ -2,12 +2,11 @@
 
 require_once ROOT_PATH . '/includes/functions.php';
 
-
 if (!isset($_GET['id'])) {
     handleError(new Exception('Missing post id'));
 }
 
-    $id = $_GET['id'];
+$id = $_GET['id'];
 
 try {
     $post = getPost($pdo,$id);
@@ -19,12 +18,19 @@ try {
     echo '<h2>' . htmlspecialchars($post['title']) . '</h2>';
     echo '<p>' . htmlspecialchars($post['content']) . '</p>';
     echo '<p>Posted by ' . htmlspecialchars($post['author']) . ' on ' . htmlspecialchars($post['date']) . '</p>';
-    echo '<a href="index.php?page=edit_post&id=' . htmlspecialchars($post['id']) . '">Edit Post</a>';
 
-    echo '<form method="POST" action="admin/delete_post.php" onsubmit="return confirm(\'Are you sure you want to delete this post?\');">';
-    echo '<input type="hidden" name="id" value="' . htmlspecialchars($post['id']) . '">';
-    echo '<input type="submit" value="Delete Post">';
-    echo '</form>';
+    if (isset($_SESSION['admin_logged_in']) && $_SESSION['admin_logged_in']) {
+        echo '<form method="GET" action="index.php">';
+        echo '<input type="hidden" name="page" value="edit_post">';
+        echo '<input type="hidden" name="id" value="' . htmlspecialchars($post['id']) . '">';
+        echo '<input type="submit" value="Edit Post">';
+        echo '</form>';
+
+        echo '<form method="POST" action="admin/delete_post.php" onsubmit="return confirm(\'Are you sure you want to delete this post?\');">';
+        echo '<input type="hidden" name="id" value="' . htmlspecialchars($post['id']) . '">';
+        echo '<input type="submit" value="Delete Post">';
+        echo '</form>';
+    }
 
     $sql = 'SELECT * FROM comments WHERE post_id = ? ORDER BY date DESC';
     $stmt = $pdo->prepare($sql);
@@ -36,21 +42,20 @@ try {
         echo '<h3>' . htmlspecialchars($comment['author']) . ' said:</h3>';
         echo '<p>' . htmlspecialchars($comment['content']) . '</p>';
         echo '<p>Posted on ' . htmlspecialchars($comment['date']) . '</p>';
-        
-        echo '<form method="POST" action="admin/delete_comment.php" onsubmit="return confirm(\'Are you sure you want to delete this comment?\');">';
-        echo '<input type="hidden" name="id" value="' . htmlspecialchars($comment['id']) . '">';
-        echo '<input type="hidden" name="post_id" value="' . htmlspecialchars($post['id']) . '">';
-        echo '<input type="submit" value="Delete Comment">';
-        echo '</form>';
+
+        if (isset($_SESSION['admin_logged_in']) && $_SESSION['admin_logged_in']) {
+            echo '<form method="POST" action="admin/delete_comment.php" onsubmit="return confirm(\'Are you sure you want to delete this comment?\');">';
+            echo '<input type="hidden" name="id" value="' . htmlspecialchars($comment['id']) . '">';
+            echo '<input type="hidden" name="post_id" value="' . htmlspecialchars($post['id']) . '">';
+            echo '<input type="submit" value="Delete Comment">';
+            echo '</form>';
+        }
     }
 } catch (PDOException $e) {
     handleError($e);
 }
 
 ?>
-
-
-
 
 <form method="post" action="public/add_comment.php">
     <input type="hidden" name="post_id" value="<?php echo htmlspecialchars($id); ?>">
@@ -60,8 +65,6 @@ try {
     <textarea id="content" name="content" required></textarea><br>
     <input type="submit" value="Submit">
 </form>
-
-
 
 <script>
 function validateForm() {
